广告系统分级查询

+++ ad-service/ad-common/src/main/java/com/youzan/ad/dump/table/UnitKeywordTable.java
@@ -13,5 +13,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UnitKeywordTable {
     private String keyword;
-    private Integer unitId;
+    private Long unitId;
 }
